Implemented basic user management, integrated management components into the app, exposed store to fetch middleware to use auth data from the store
